+++ qiskit/providers/aer/noise/errors/__init__.py
@@ -7,8 +7,8 @@
 
 """Standard error module for Qiskit Aer."""
 
-from .quantum_error import QuantumError
 from .readout_error import ReadoutError
+from .quantum_error import QuantumError
 from .standard_errors import kraus_error
 from .standard_errors import mixed_unitary_error
 from .standard_errors import coherent_unitary_error
@@ -19,3 +19,4 @@ from .standard_errors import thermal_relaxation_error
 from .standard_errors import phase_amplitude_damping_error
 from .standard_errors import amplitude_damping_error
 from .standard_errors import phase_damping_error
+